requirements.txt
streamlit==1.32.0
anthropic==0.18.1
python-dotenv==1.0.0
